+++ fidget/src/render/render2d.rs
@@ -9,16 +9,23 @@ use nalgebra::Point2;
 
 ////////////////////////////////////////////////////////////////////////////////
 
+/// Response type for [`RenderMode::interval`]
+pub enum IntervalAction<T> {
+    Fill(T),
+    Interpolate,
+    Recurse,
+}
+
 /// Configuration trait for rendering
 pub trait RenderMode {
     /// Type of output pixel
     type Output: Default + Copy + Clone + Send;
 
     /// Decide whether to subdivide or fill an interval
-    fn interval(&self, i: Interval, depth: usize) -> Option<Self::Output>;
+    fn interval(i: Interval, depth: usize) -> IntervalAction<Self::Output>;
 
     /// Per-pixel drawing
-    fn pixel(&self, f: f32) -> Self::Output;
+    fn pixel(f: f32) -> Self::Output;
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -28,24 +35,24 @@ pub struct DebugRenderMode;
 
 impl RenderMode for DebugRenderMode {
     type Output = DebugPixel;
-    fn interval(&self, i: Interval, depth: usize) -> Option<DebugPixel> {
+    fn interval(i: Interval, depth: usize) -> IntervalAction<DebugPixel> {
         if i.upper() < 0.0 {
             if depth > 1 {
-                Some(DebugPixel::FilledSubtile)
+                IntervalAction::Fill(DebugPixel::FilledSubtile)
             } else {
-                Some(DebugPixel::FilledTile)
+                IntervalAction::Fill(DebugPixel::FilledTile)
             }
         } else if i.lower() > 0.0 {
             if depth > 1 {
-                Some(DebugPixel::EmptySubtile)
+                IntervalAction::Fill(DebugPixel::EmptySubtile)
             } else {
-                Some(DebugPixel::EmptyTile)
+                IntervalAction::Fill(DebugPixel::EmptyTile)
             }
         } else {
-            None
+            IntervalAction::Recurse
         }
     }
-    fn pixel(&self, f: f32) -> DebugPixel {
+    fn pixel(f: f32) -> DebugPixel {
         if f < 0.0 {
             DebugPixel::Filled
         } else {
@@ -101,29 +108,33 @@ pub struct BitRenderMode;
 
 impl RenderMode for BitRenderMode {
     type Output = bool;
-    fn interval(&self, i: Interval, _depth: usize) -> Option<bool> {
+    fn interval(i: Interval, _depth: usize) -> IntervalAction<bool> {
         if i.upper() < 0.0 {
-            Some(true)
+            IntervalAction::Fill(true)
         } else if i.lower() > 0.0 {
-            Some(false)
+            IntervalAction::Fill(false)
         } else {
-            None
+            IntervalAction::Recurse
         }
     }
-    fn pixel(&self, f: f32) -> bool {
+    fn pixel(f: f32) -> bool {
         f < 0.0
     }
 }
 
-/// Rendering mode which mimicks many SDF demos on ShaderToy
-pub struct SdfRenderMode;
+/// Pixel-perfect render mode which mimicks many SDF demos on ShaderToy
+///
+/// This mode recurses down to individual pixels, so it doesn't take advantage
+/// of skipping empty / full regions; use [`SdfRenderMode`] for a
+/// faster-but-approximate visualization.
+pub struct SdfPixelRenderMode;
 
-impl RenderMode for SdfRenderMode {
+impl RenderMode for SdfPixelRenderMode {
     type Output = [u8; 3];
-    fn interval(&self, _i: Interval, _depth: usize) -> Option<[u8; 3]> {
-        None // always recurse
+    fn interval(_i: Interval, _depth: usize) -> IntervalAction<[u8; 3]> {
+        IntervalAction::Recurse
     }
-    fn pixel(&self, f: f32) -> [u8; 3] {
+    fn pixel(f: f32) -> [u8; 3] {
         let r = 1.0 - 0.1f32.copysign(f);
         let g = 1.0 - 0.4f32.copysign(f);
         let b = 1.0 - 0.7f32.copysign(f);
@@ -145,6 +156,26 @@ impl RenderMode for SdfRenderMode {
         };
 
         [run(r), run(g), run(b)]
+    }
+}
+
+/// Fast rendering mode which mimicks many SDF demos on ShaderToy
+///
+/// Unlike [`SdfPixelRenderMode`], this mode uses linear interpolation when
+/// evaluating empty or full regions, which is significantly faster.
+pub struct SdfRenderMode;
+
+impl RenderMode for SdfRenderMode {
+    type Output = [u8; 3];
+    fn interval(i: Interval, _depth: usize) -> IntervalAction<[u8; 3]> {
+        if i.upper() < 0.0 || i.lower() > 0.0 {
+            IntervalAction::Interpolate
+        } else {
+            IntervalAction::Recurse
+        }
+    }
+    fn pixel(f: f32) -> [u8; 3] {
+        SdfPixelRenderMode::pixel(f)
     }
 }
 
@@ -194,7 +225,6 @@ impl<S: Shape, M: RenderMode> Worker<'_, S, M> {
         shape: &mut RenderHandle<S>,
         depth: usize,
         tile: Tile<2>,
-        mode: &M,
     ) {
         let tile_size = self.config.tile_sizes[depth];
 
@@ -209,14 +239,43 @@ impl<S: Shape, M: RenderMode> Worker<'_, S, M> {
             .eval(shape.i_tape(&mut self.tape_storage), x, y, z)
             .unwrap();
 
-        let fill = mode.interval(i, depth);
-
-        if let Some(fill) = fill {
-            for y in 0..tile_size {
-                let start = self.config.tile_to_offset(tile, 0, y);
-                self.image[start..][..tile_size].fill(fill);
+        match M::interval(i, depth) {
+            IntervalAction::Fill(fill) => {
+                for y in 0..tile_size {
+                    let start = self.config.tile_to_offset(tile, 0, y);
+                    self.image[start..][..tile_size].fill(fill);
+                }
+                return;
             }
-            return;
+            IntervalAction::Interpolate => {
+                let xs = [x.lower(), x.lower(), x.upper(), x.upper()];
+                let ys = [y.lower(), y.upper(), y.lower(), y.upper()];
+                let zs = [0.0; 4];
+                let vs = self
+                    .eval_float_slice
+                    .eval(shape.f_tape(&mut self.tape_storage), &xs, &ys, &zs)
+                    .unwrap();
+                // Bilinear interpolation on a per-pixel basis
+                for y in 0..tile_size {
+                    // Y interpolation
+                    let y_frac = (y as f32 - 1.0) / (tile_size as f32);
+                    let v0 = vs[0] * (1.0 - y_frac) + vs[1] * y_frac;
+                    let v1 = vs[2] * (1.0 - y_frac) + vs[3] * y_frac;
+
+                    let mut i = self.config.tile_to_offset(tile, 0, y);
+                    for x in 0..tile_size {
+                        // X interpolation
+                        let x_frac = (x as f32 - 1.0) / (tile_size as f32);
+                        let v = v0 * (1.0 - x_frac) + v1 * x_frac;
+
+                        // Write out the pixel
+                        self.image[i] = M::pixel(v);
+                        i += 1;
+                    }
+                }
+                return;
+            }
+            IntervalAction::Recurse => (), // keep going
         }
 
         let sub_tape = if let Some(trace) = simplify.as_ref() {
@@ -241,12 +300,11 @@ impl<S: Shape, M: RenderMode> Worker<'_, S, M> {
                             tile.corner[0] + i * next_tile_size,
                             tile.corner[1] + j * next_tile_size,
                         ]),
-                        mode,
                     );
                 }
             }
         } else {
-            self.render_tile_pixels(sub_tape, tile_size, tile, mode);
+            self.render_tile_pixels(sub_tape, tile_size, tile);
         }
     }
 
@@ -255,7 +313,6 @@ impl<S: Shape, M: RenderMode> Worker<'_, S, M> {
         shape: &mut RenderHandle<S>,
         tile_size: usize,
         tile: Tile<2>,
-        mode: &M,
     ) {
         let mut index = 0;
         for j in 0..tile_size {
@@ -280,7 +337,7 @@ impl<S: Shape, M: RenderMode> Worker<'_, S, M> {
         for j in 0..tile_size {
             let o = self.config.tile_to_offset(tile, 0, j);
             for i in 0..tile_size {
-                self.image[o + i] = mode.pixel(out[index]);
+                self.image[o + i] = M::pixel(out[index]);
                 index += 1;
             }
         }
@@ -293,7 +350,6 @@ fn worker<S: Shape, M: RenderMode>(
     mut shape: RenderHandle<S>,
     queue: &Queue<2>,
     config: &AlignedRenderConfig<2>,
-    mode: &M,
 ) -> Vec<(Tile<2>, Vec<M::Output>)> {
     let mut out = vec![];
     let scratch = Scratch::new(config.tile_sizes.last().unwrap_or(&0).pow(2));
@@ -310,7 +366,7 @@ fn worker<S: Shape, M: RenderMode>(
     };
     while let Some(tile) = queue.next() {
         w.image = vec![M::Output::default(); config.tile_sizes[0].pow(2)];
-        w.render_tile_recurse(&mut shape, 0, tile, mode);
+        w.render_tile_recurse(&mut shape, 0, tile);
         let pixels = std::mem::take(&mut w.image);
         out.push((tile, pixels))
     }
@@ -331,7 +387,6 @@ fn worker<S: Shape, M: RenderMode>(
 pub fn render<S: Shape, M: RenderMode + Sync>(
     shape: S,
     config: &RenderConfig<2>,
-    mode: &M,
 ) -> Vec<M::Output> {
     let (config, mat) = config.align();
     assert!(config.image_size % config.tile_sizes[0] == 0);
@@ -344,13 +399,12 @@ pub fn render<S: Shape, M: RenderMode + Sync>(
     let mat = mat.insert_column(2, 0.0);
     let shape = shape.apply_transform(mat);
 
-    render_inner(shape, config, mode)
+    render_inner::<_, M>(shape, config)
 }
 
 fn render_inner<S: Shape, M: RenderMode + Sync>(
     shape: S,
     config: AlignedRenderConfig<2>,
-    mode: &M,
 ) -> Vec<M::Output> {
     let mut tiles = vec![];
     for i in 0..config.image_size / config.tile_sizes[0] {
@@ -369,9 +423,7 @@ fn render_inner<S: Shape, M: RenderMode + Sync>(
     let _ = rh.i_tape(&mut vec![]); // populate i_tape before cloning
 
     let out: Vec<_> = if threads == 1 {
-        worker::<S, M>(rh, &queue, &config, mode)
-            .into_iter()
-            .collect()
+        worker::<S, M>(rh, &queue, &config).into_iter().collect()
     } else {
         #[cfg(target_arch = "wasm32")]
         unreachable!("multithreaded rendering is not supported on wasm32");
@@ -381,9 +433,7 @@ fn render_inner<S: Shape, M: RenderMode + Sync>(
             let mut handles = vec![];
             for _ in 0..threads {
                 let rh = rh.clone();
-                handles.push(
-                    s.spawn(|| worker::<S, M>(rh, &queue, &config, mode)),
-                );
+                handles.push(s.spawn(|| worker::<S, M>(rh, &queue, &config)));
             }
             let mut out = vec![];
             for h in handles {
@@ -440,7 +490,7 @@ mod test {
             bounds,
             ..RenderConfig::default()
         };
-        let out = cfg.run(shape, &BitRenderMode).unwrap();
+        let out = cfg.run::<_, BitRenderMode>(shape).unwrap();
         let mut img_str = String::new();
         for (i, b) in out.iter().enumerate() {
             if i % 32 == 0 {